Completed word detail tests

+++ src/Domain/Commands/UpdateWordDetailCommand.cs
@@ -4,7 +4,7 @@ namespace Inshapardaz.Domain.Commands
 {
     public class UpdateWordDetailCommand : Command
     {
-        public int WordId { get; set; }
+        public long WordId { get; set; }
 
         public WordDetail WordDetail { get; set; }
     }
